Changes for pressure balance.
Ths version of the script 1) First interpolates density and Br, and only then computes cs. 2) Computes cs based on pressure balance.

There is also a minor change that sets the number of iterations in the Krylov solver, but this is most probably temporary (was used for higher resolution calculation, but that does not work anyway.